==> src/2.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_json("article_metadata.json")

df.head()

df["PROPN"] = df["pos_ratios"].apply(lambda x: x.get("PROPN", 0))
df["ADV"] = df["pos_ratios"].apply(lambda x: x.get("ADV", 0))
df["VERB"] = df["pos_ratios"].apply(lambda x: x.get("VERB", 0))
df["DET"] = df["pos_ratios"].apply(lambda x: x.get("DET", 0))
df["CCONJ"] = df["pos_ratios"].apply(lambda x: x.get("CCONJ", 0))
df["PRON"] = df["pos_ratios"].apply(lambda x: x.get("PRON", 0))
df["ADP"] = df["pos_ratios"].apply(lambda x: x.get("ADP", 0))
df["PART"] = df["pos_ratios"].apply(lambda x: x.get("PART", 0))
df["NOUN"] = df["pos_ratios"].apply(lambda x: x.get("NOUN", 0))
df["ADJ"] = df["pos_ratios"].apply(lambda x: x.get("ADJ", 0))
df["NUM"] = df["pos_ratios"].apply(lambda x: x.get("NUM", 0))
df["SCONJ"] = df["pos_ratios"].apply(lambda x: x.get("SCONJ", 0))
df["AUX"] = df["pos_ratios"].apply(lambda x: x.get("AUX", 0))

df = df.drop(["pos_ratios"], axis=1)


df.head()

df.to_csv('final.csv')
