==> src/3.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv("final.csv")

df.head()

df["has_clickbait_words"] = df["has_clickbait_words"].apply(lambda x: 0 if x == "False" else 1)


df.head()

df = df.drop(["id","subject","date"],axis=1)

df.head()

df = df.drop(["Unnamed: 0"],axis=1)

df.columns

from sklearn.preprocessing import MinMaxScaler

scaler = MinMaxScaler()
df[['num_characters', 'num_words', 'num_sentences', 'avg_sentence_length',
       'num_capitalized_words', 'num_exclamations', 'num_questions',
       'has_clickbait_words', 'readability_score', 'sentiment_polarity',
       'sentiment_subjectivity', 'PROPN', 'ADV', 'VERB', 'DET',
       'CCONJ', 'PRON', 'ADP', 'PART', 'NOUN', 'ADJ', 'NUM', 'SCONJ', 'AUX']] = scaler.fit_transform(df[['num_characters', 'num_words', 'num_sentences', 'avg_sentence_length',
       'num_capitalized_words', 'num_exclamations', 'num_questions',
       'has_clickbait_words', 'readability_score', 'sentiment_polarity',
       'sentiment_subjectivity', 'PROPN', 'ADV', 'VERB', 'DET',
       'CCONJ', 'PRON', 'ADP', 'PART', 'NOUN', 'ADJ', 'NUM', 'SCONJ', 'AUX']])


df

df.to_csv("Data.csv")
