==> src/4.ipynb <==
import tensorflow as tf
from tensorflow.keras.models import Sequential
from tensorflow.keras.layers import Dense
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler
import pandas as pd

df = pd.read_csv("../Data/Data.csv")

df = df.drop(["Unnamed: 0"],axis=1)

print(df.head().to_string())

X = df.drop("class", axis=1) 
y = df["class"]                


X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

scaler = StandardScaler()
X_train_scaled = scaler.fit_transform(X_train)
X_test_scaled = scaler.transform(X_test)

model = Sequential([
    Dense(256, activation='relu', input_shape=(24,)),
    Dense(128, activation='relu'),
    Dense(64, activation='relu'),  
    Dense(32, activation='relu'),
    Dense(16, activation='relu'),                      
    Dense(1, activation='sigmoid')                     
])


model.summary()

model.compile(optimizer='adam',loss='binary_crossentropy',metrics=['accuracy'])


from tensorflow.keras.callbacks import EarlyStopping


early_stop = EarlyStopping(
    monitor='val_accuracy',      
    patience=30,               
    restore_best_weights=True  
)

history = model.fit(
    X_train_scaled, y_train,
    epochs=1000,
    batch_size=32,
    validation_data=(X_test_scaled, y_test),
    callbacks=[early_stop]
)


test_loss, test_accuracy = model.evaluate(X_test_scaled, y_test)
print(f"Test Accuracy: {test_accuracy:.4f}")
